Placed business logic in viewmodel, and ui logic in mainscreen.
